date format change & initial wish load optimisation
Changed the month display from number to shortened text like "Jan". Also
the server no longer keeps all the users in array and when they request
data we look them up via ID in the array. Currently we just respond to
the one who sent the request. I did that in order to avoid adding
scrubber for the offline users.